feat: 0.5.0 consumer contract for ratebook evaluation, bounds and fail-loud outputs
Specified in docs/DESIGN_DECISIONS.md section 13.

- One canonical Rust kernel (evaluate_ratebook) evaluates the final factor
  tables per quote; every reported ratebook total comes from it.
  RatebookResult gains quote_results, clamp counts, scenario_values,
  baseline_scenario_value and constraint_bounds; RatebookOptimiser.evaluate()
  reproduces a result exactly.
- The ratebook frontier keeps each point's factor tables (a row's totals are
  their canonical evaluation) and releases per-point results as it goes.
- Every frontier emits bound_<c>; solve results expose constraint_bounds.
- One baseline rule (nearest 1.0 in f32) for sum, pct and ratio constraints;
  ratio totals accumulate in f64.
- Fail loud: unknown warm-start lambda names, zero-baseline pct (sum and
  ratio), non-positive candidates, non-strictly-increasing scenario values,
  reserved constraint names, ambiguous factor names and separators, ignored
  parallel=True, and missing frontier/summary values all raise.
- Ordered dict outputs, explicit PerFactorRecord fields, save format 2 with
  ResultUnavailableError, quote_results_schema/frontier_points_schema,
  py.typed and updated stubs.

BREAKING CHANGE: RatebookResult, frontier point columns, per_factor_results,
the save format and several newly raised errors change the public contract.

+++ python/price_contour/_ratio_results.py
@@ -44,8 +44,10 @@ def _safe_ratio_from_columns(
     output, surfacing the divide-by-zero loud rather than silently
     reporting zero.
     """
-    num_total = float(df[numerator_col].sum())
-    denom_total = float(df[denominator_col].sum())
+    # Float64 accumulation of the (Float32) values, like every other total,
+    # so a reported ratio agrees with the ratio bound derived from it.
+    num_total = float(df[numerator_col].cast(pl.Float64).sum())
+    denom_total = float(df[denominator_col].cast(pl.Float64).sum())
     if denom_total == 0.0:
         return float("nan")
     return num_total / denom_total
